requirements.txt
asgiref==3.5.0
Django==3.2
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
